+++ CHANGES.md
@@ -39,6 +39,7 @@ Changes in 12.5
 - JS remove inline code from RegistrationSiblingUseContactsAddress() in Registration.fnc.php & RegistrationSiblingUseContactsAddress.js
 - JS remove inline code from program in MassCreateAssignments.php & MassCreateAssignments.js
 - JS remove inline code from StudentAssignmentSubmissionOutput() in StudentAssignments.fnc.php & StudentAssignmentSubmissionOutput.js
+- JS remove inline code from program in ChooseCourse.php & ChooseCourse.js
 
 Changes in 12.4.3
 -----------------